+++ ChilledCafe/Views/Main/TabBarIcon.swift
@@ -1,0 +1,27 @@
+//
+//  TabbarIcon.swift
+//  ChilledCafe
+//
+//  Created by Terry Koo on 2022/12/06.
+//
+import SwiftUI
+
+struct TabBarIcon: View {
+    @StateObject var viewRouter: ViewRouter
+    let assignedPage: Page
+    let width, height: CGFloat
+    let iconName: String
+
+    var body: some View {
+        VStack {
+            Image(iconName)
+                .resizable()
+                .aspectRatio(contentMode: .fit)
+                .frame(width: width, height: height)
+        }
+            .onTapGesture {
+                viewRouter.currentPage = assignedPage
+            }
+            .foregroundColor(viewRouter.currentPage == assignedPage ? .blue : .gray)
+    }
+}
